config/db.py
from decouple import config

MYSQL = {
    'default': {
        'ENGINE': 'django.db.backends.mysql',
        'NAME': config('DB_NAME'),
        'USER': config('DB_USER'),
        'PASSWORD': config('DB_PASSWORD'),
        'PORT': 3306,
        'HOST': 'localhost',
    }
}