requirements.txt
APScheduler==3.9.1.post1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.2
autopep8==1.5.7
aniso8601==7.0.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
charset-normalizer==2.0.12
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.8
django-filebrowser-no-grappelli==3.8.0
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
django-allauth==0.50.0
django-appconf==1.0.5
django-apscheduler==0.6.2
django-extensions==3.1.5
django-filer==2.1.2
django-htmx==1.6.0
django-jazzmin==2.5.0
django-polymorphic==3.0.0
drf-yasg==1.20.0
filelock==3.0.12
fontawesomefree==6.0.0
future==0.18.2
graphene
graphene-django<=2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
httplib2==0.20.4
idna==3.3
ipython==7.28.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsmin==2.2.2
lxml==4.7.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.26
oauth2client==4.1.3
oauthlib==3.2.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.23
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
packaging==20.7
promise==2.3
pyenchant==3.1.1
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==1.6.1
sqlparse==0.4.2
singledispatch==3.4.0.3
six==1.15.0
tzdata==2021.5
tzlocal==4.1
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.7.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.6
xlrd==2.0.1
xlwt==1.3.0